added exclusions and added regexp for processing files referred by requirejs

+++ tasks/cachebust.js
@@ -9,6 +9,10 @@ module.exports = function(grunt) {
             src: /<script.+src=['"](?!http:|https:|\/\/)([^"']+)["']/gm,
             file: /src=['"]([^"']+)["']/m
         },
+		requirejs: {
+            src: /<script.+data-main=['"](?!http:|https:|\/\/)([^"']+)["']/gm,
+            file: /data-main=['"]([^"']+)["']/m
+		},
         css: {
             src: /<link.+href=["'](?!http:|https:|\/\/).*\.css("|\?.*")/gm,
             file: /href=['"]([^"']+)["']/m
@@ -32,10 +36,12 @@ module.exports = function(grunt) {
         replaceTerms:[],
         rename: false
     };
+	
 
     grunt.registerMultiTask('cacheBust', 'Bust static assets from the cache using content hashing', function() {
 
         var opts = grunt.util._.defaults(this.options(), options);
+		var excluded = this.data.exclusions || [];
 
         this.files.forEach(function(f) {
             var src = f.src.filter(function(filepath) {
@@ -44,54 +50,55 @@ module.exports = function(grunt) {
                     grunt.log.warn('Source file "' + filepath + '" not found.');
                     return false;
                 } else {
-                    return true;
+					return true;
                 }
             }).map(function(filepath) {
                 var data = grunt.file.read(filepath, fileOptions);
-
                 grunt.util._.each(regexs, function(regex, type) {
                     var matches = data.match(regex.src) || [];
                     matches.forEach(function(snippet) {
+						// Don't need to generate hash for excluded files
+						if (excluded.indexOf(snippet.match(regex.file)[1]) === -1) {
+							// Generate hash
+							var hash = opts.hash || crypto.createHash(opts.algorithm).update(data, opts.encoding).digest('hex').substring(0, opts.length);
 
-                        // Generate hash
-                        var hash = opts.hash || crypto.createHash(opts.algorithm).update(data, opts.encoding).digest('hex').substring(0, opts.length);
+							var extension = type !== 'images' ? '.'+ type : snippet.match(/\.\w+/)[0];
 
-                        var extension = type !== 'images' ? '.'+ type : snippet.match(/\.\w+/)[0];
+							if(opts.rename) {
+								var path     = opts.baseDir + '/';
+								var _snippet = snippet;
 
-                        if(opts.rename) {
-                            var path     = opts.baseDir + '/';
-                            var _snippet = snippet;
+								// Replacing specific terms in the import path so renaming files
+								if(opts.replaceTerms && opts.replaceTerms.length > 0) {
+									opts.replaceTerms.forEach(function(obj) {
+										grunt.util._.each(obj, function(replacement, term) {
+											snippet = snippet.replace(term, replacement);
+										});
+									});
+								}
 
-                            // Replacing specific terms in the import path so renaming files
-                            if(opts.replaceTerms && opts.replaceTerms.length > 0) {
-                                opts.replaceTerms.forEach(function(obj) {
-                                    grunt.util._.each(obj, function(replacement, term) {
-                                        snippet = snippet.replace(term, replacement);
-                                    });
-                                });
-                            }
+								// Remove duplicate hashes
+								snippet = snippet.replace(new RegExp('_'+ opts.hash+'|[a-zA-Z0-9]{'+ opts.length +'}', 'ig'), '');
+								_snippet = _snippet.replace(new RegExp('_'+ opts.hash+'|[a-zA-Z0-9]{'+ opts.length +'}', 'ig'), '');
 
-                            // Remove duplicate hashes
-                            snippet = snippet.replace(new RegExp('_'+ opts.hash+'|[a-zA-Z0-9]{'+ opts.length +'}', 'ig'), '');
-                            _snippet = _snippet.replace(new RegExp('_'+ opts.hash+'|[a-zA-Z0-9]{'+ opts.length +'}', 'ig'), '');
-
-                            var name = snippet.match(regex.file)[1];
-                            var filename    = path + name;
-                            var newFilename = path + name.replace(extension, '') +'_'+ hash + extension;
+								var name = snippet.match(regex.file)[1];
+								var filename    = path + name;
+								var newFilename = path + name.replace(extension, '') +'_'+ hash + extension;
 
 
-                            _snippet = _snippet.substring(0, _snippet.length - 1);
-                            data     = data.replace(_snippet, _snippet.replace(extension, '') +'_'+ hash + extension);
+								_snippet = _snippet.substring(0, _snippet.length - 1);
+								data     = data.replace(_snippet, _snippet.replace(extension, '') +'_'+ hash + extension);
 
-                            grunt.file.copy(filename, newFilename);
+								grunt.file.copy(filename, newFilename);
 
-                            if(opts.deleteOriginals) {
-                                grunt.file.delete(filename);
-                            }
-                        } else {
-                            snippet = snippet.substring(0, snippet.length - 1);
-                            data    = data.replace(snippet, snippet.split('?')[0] + '?' + hash);
-                        }
+								if(opts.deleteOriginals) {
+									grunt.file.delete(filename);
+								}
+							} else {
+								snippet = snippet.substring(0, snippet.length - 1);
+								data    = data.replace(snippet, snippet.split('?')[0] + '?' + hash);
+							}
+						}
                     });
                 });
